requirements.txt
PyQt5
PyQt5-sip
PuLP
google-generativeai
numpy
scipy
networkx
matplotlib